+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Integrated Vesktop Flatpak support
+
 ## [1.7.3] - 2025-01-25
 
 ### Changed